#118
When getValueFromJS generate Exception (due to bad execution of JS)
Or when the value calculated is null (not "null" string but null value of string)

Generate a KO on calculate property and send a generic message.

In order to get a KO on testcase execution and not a "bad" pending.