Deploy binaries to a SourceForge repository
When using the ``release`` command (from ``tools/release/release.[sh|bat]``) then release binaries are deployed to a SourceForge repository.

In order to be able to do the release, then you have to fill your ``~.m2/settings.xml`` file with the following information:

```
  <profiles>
    <profile>
      <id>default</id>
      <properties>
        <sourceforge.username>USERNAME,cerberus-source</sourceforge.username>
        <sourceforge.password>PASSWORD</sourceforge.password>
        <sourceforge.url>frs.sourceforge.net:/home/frs/project/cerberus-source/</sourceforge.url>
      </properties>
    </profile>
  </profiles>

  <activeProfiles>
    <activeProfile>default</activeProfile>
  </activeProfiles>
```

Theses lines create a new profile, named ``default``, that fill the ``sourceforge.**`` variables which are required for the deploy process (see the ``source/pom.xml`` for more details).

Closes #398