#1814 first implementation
POC for secondary properties.

TestCaseExecution page implementation.

How to: add '[secondary]' to the property description and save it in order to make it appears as a secondary.

Goal: have a better readability on some properties that can be less significant to the business side and allows to focus on the most relevant ones.