#2105 First commit for TestCase model update
This commit will break a considerable part of the service. The JSON returned by TestCase is updated with the last keys decided. Only one attribute has been updated in this commit : group -> type.

Still a lot of work to update the string parameters in the different classes

Datababase model will be updated in next commit